[ISSUE-196] Remove some unused code

+++ angel-ps/core/src/main/java/com/tencent/angel/psagent/matrix/oplog/cache/LongKeySparseDoubleRowUpdateSplit.java
@@ -46,7 +46,7 @@ public class LongKeySparseDoubleRowUpdateSplit extends RowUpdateSplit{
   @Override
   public void serialize(ByteBuf buf) {
     super.serialize(buf);
-    buf.writeInt((int)(end - start));
+    buf.writeInt(end - start);
     LOG.debug("double size = " + (end - start));
     for (int i = start; i < end; i++) {
       buf.writeLong(offsets[i]);
