+++ angel-ps/core/src/main/java/com/tencent/angel/psagent/matrix/oplog/cache/SparseIntRowUpdateSplit.java
@@ -64,7 +64,7 @@ public class SparseIntRowUpdateSplit extends RowUpdateSplit {
   @Override
   public void serialize(ByteBuf buf) {
     super.serialize(buf);
-    buf.writeInt((int)(end - start));
+    buf.writeInt(end - start);
     LOG.debug("int size = " + (end - start));
     for (int i = start; i < end; i++) {
       buf.writeInt(offsets[i]);
